OddEven, OnewTwoALot solved, AverageOfInputs int variables changed to double

+++ week-01/day-04/src/average_of_input/AverageOfInput.java
@@ -6,11 +6,11 @@ public class AverageOfInput {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.println("Give me 5 numbers to get their average");
-        int num1 = sc.nextInt();
-        int num2 = sc.nextInt();
-        int num3 = sc.nextInt();
-        int num4 = sc.nextInt();
-        int num5 = sc.nextInt();
+        double num1 = sc.nextDouble();
+        double num2 = sc.nextDouble();
+        double num3 = sc.nextDouble();
+        double num4 = sc.nextDouble();
+        double num5 = sc.nextDouble();
 
         double avg = (num1 + num2 + num3 + num4 + num5) / 5;
